Moved state management for date to a custom hook
