+++ src/App.tsx
@@ -1,22 +1,12 @@
-import { useState, useCallback, useMemo } from "react";
-import type { ChangeEvent } from "react";
-
 import Map from "./Map";
 
 import travels from "./travels";
 import DateSlider from "./DateSlider";
 
+import useDate from "./useDate";
+
 const App = () => {
-  const [date, setDate] = useState(travels[0].date.getTime());
-
-  const onHandleDateChange = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => {
-      setDate(Number(event.target.value));
-    },
-    []
-  );
-
-  const firstTravelDate = useMemo(() => travels[0].date.getTime(), []);
+  const { date, onHandleDateChange, firstTravelDate } = useDate();
 
   return (
     <>
